fix(parser): fix the header regex pattern
We should consider `foo(bar)` format in parsing the
header. The previous header pattern, which is defined in
conventional-changelog-angular package, assumes `:` character
always exists in the headers with  `type(scope): subject`
format, but since the subject is optional, the colon
character might not exist and still we have type and scope
sections.